fix: the audio path stops allocating, and three smaller things the review named
The read path rented one buffer per session instead of a MemoryStream and an
8 kB chunk per message, and the jitter buffer rents its sample arrays from
the pool instead of allocating one per block. That was about 25 kB of
garbage per 40 ms block per watched station, 600 kB/s each, on exactly the
kind of receive path CLAUDE.md's zero-steady-state-allocation rule is about.
The station's own client already pools the identical path. The samples now
go from the socket's bytes into the buffer with nothing allocated in
between.

A relayed frame's own timestamp is held to a day either side of this site's
clock. A station could date a frame in the year 9999 and it was written into
the site's copy of its log and sorted above everything else on its page for
ever. Self-harm rather than an attack, and now treated like every other
untrusted field here. A plausible time is still kept exactly, because it is
the station's own and better than the moment its bytes crossed the wire.

A bearer value this site could never have issued - empty, or longer than any
token it hands out - is delayed and counted like any other guess. It used to
fall into the "no token at all" arm and be refused instantly and silently,
so a run of them was invisible in the journal.

And the tail of a run of bad tokens is no longer lost. The line carries the
count up to it, so a run that stopped before the next line was due ended
with the journal reporting the first attempt and never mentioning the other
two hundred. A timer closes the window and says whatever the run reached.

A_Station_Cannot_Choose_Its_Own_Slug now sends the slug on the first and
only hello, which is the one that is read: on a second hello it was never
looked at, because a second hello ends the session, so the test passed off
the first welcome and proved nothing about the field.

+++ src/Packet.SoundModem.Daemon/RelayStation.cs
@@ -575,8 +575,8 @@ internal sealed class RelayStation : IMonitorStation
         live?.Demand(viewers);
     }
 
-    /// <summary>Audio off the wire.</summary>
-    internal void PushAudio(ReadOnlySpan<short> pcm, bool transmitted) =>
+    /// <summary>Audio off the wire, as the little-endian bytes it arrived as.</summary>
+    internal void PushAudio(ReadOnlySpan<byte> pcm, bool transmitted) =>
         _input.Push(pcm, transmitted);
 
     /// <summary>The station's own sentence about its radio, for the page's status chip.</summary>
